requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
PyPDF2
python-multipart
python-docx
sqlalchemy
psycopg2
alembic
python-dotenv
pytz
bcrypt
pyjwt
pydantic[email]
fastapi-mail